requirements.txt
# Core dependencies for the Discord bot and async operations
discord.py==2.3.2
aiohttp==3.9.5
aiomysql==0.2.0
redis==5.0.4
cryptography==42.0.5
aiohttp-session==2.12.0
jinja2==3.1.2
python-dotenv==1.0.1
Pillow==10.3.0
PyTZ==2024.1
typing_extensions==4.11.0
jsonpickle==3.0.4
regex==2024.5.15
pandas==2.2.2
apscheduler==3.10.4